insights_agent/requirements.txt
a2a-sdk[http-server]>=0.2.5
uvicorn>=0.27.0
requests>=2.31.0
cryptography>=42.0.0
pyjwt>=2.8.0
python-dotenv>=1.0.1
